New features
- Use flex-grow property for simple grids by using grid-cell--light and grid-cell--heavy (inspired from cutestrap grid)
- Add grid--auto and grid--auto-{i} classes to grid to avoid having to define "grid-cell" on each child div. This makes simple grids really easy to make without having tons of markup

This is not fully tested but seems fine on my current project.

I'd be happy to update the docs if you like the changes. I'd also like to find a way to use negative margins instead of padding because it makes styling much easier without extra divs to wrap the content.